app/src/main/java/lulian/ballinbistrobuddies/view/LunchMenuView.java
package lulian.ballinbistrobuddies.view;

import android.os.Bundle;

import android.webkit.WebSettings;
import android.webkit.WebView;
import androidx.activity.EdgeToEdge;
import androidx.appcompat.app.AppCompatActivity;
import androidx.core.graphics.Insets;
import androidx.core.view.ViewCompat;
import androidx.core.view.WindowInsetsCompat;

import lulian.ballinbistrobuddies.R;

/**
 * LunchMenuView is an activity class that displays a web page in a WebView.
 * This class is designed to show the lunch menu from a specific URL. It demonstrates
 * the use of WebView to load web content, and the application of edge-to-edge display
 * to make the content immersive by handling system bars.
 */
public class LunchMenuView extends AppCompatActivity {

    private WebView webView;

    /**
     * Called when the activity is starting. This is where most initialization should go:
     * calling setContentView(int) to inflate the activity's UI, using findViewById(int)
     * to programmatically interact with widgets in the UI, setting up listeners, and so on.
     *
     * @param savedInstanceState If the activity is being re-initialized after previously being shut down,
     *                           this Bundle contains the data it most recently supplied in onSaveInstanceState(Bundle).
     *                           Otherwise, it is null.
     */
    @Override
    protected void onCreate(Bundle savedInstanceState) {
        super.onCreate(savedInstanceState);

        // Enable edge-to-edge display
        EdgeToEdge.enable(this);

        // Set the layout for this activity
        setContentView(R.layout.activity_lunch_menu_view);

        // Initialize WebView from layout
        webView = findViewById(R.id.webView);

        // Apply window insets to handle system bars (status bar and navigation bar)
        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
            return insets;
        });

        // Configure WebView settings
        WebSettings webSettings = webView.getSettings();
        webSettings.setJavaScriptEnabled(true); // Enable JavaScript (optional, depending on the website)

        // Load a web page into WebView
        webView.loadUrl("https://www.stw.berlin/mensen/einrichtungen/hochschule-f%C3%BCr-technik-und-wirtschaft-berlin/mensa-htw-wilhelminenhof.html");
    }
}